refactor

+++ events/messageDelete.ts
@@ -1,9 +1,26 @@
-import { Events } from "discord.js";
+import { Events, Message } from "discord.js";
 import { prisma } from "../utilities/db";
 
 const name = Events.MessageDelete;
 
-const execute = async (message) => {
+const execute = async (message: Message) => {
+    if (!message.guild || message.channel.isDMBased()) return;
+    await prisma.guild.upsert({
+        where: {
+            id: message.guild.id,
+        },
+        create: {
+            id: message.guild.id,
+        },
+        update: {},
+    });
+    const guild = await prisma.guild.findUnique({
+        where: {
+            id: message.guild.id,
+        },
+    });
+    if (!guild || !guild.logging) return;
+
     try {
         await prisma.message.delete({ where: { id: message.id } });
     } catch (err) {
